+++ include/bbque/app/working_mode_status.h
@@ -45,7 +45,7 @@ typedef std::shared_ptr<Application> AppPtr_t;
 /**  Shared pointer to TransitionOverheads */
 typedef std::shared_ptr<TransitionOverheads> OverheadPtr_t;
 /** Map of OverheadPtr_t. Key: destination working mode name */
-typedef std::map<uint16_t, OverheadPtr_t> OverheadsMap_t;
+typedef std::map<uint8_t, OverheadPtr_t> OverheadsMap_t;
 /** Bitset for clusters use monitoring */
 typedef std::bitset<MAX_NUM_CLUSTERS> ClustersBitSet;
 
@@ -86,7 +86,7 @@ public:
 	 * @brief Working Mode ID
 	 * @return ID number
 	 */
-	virtual uint16_t Id() const = 0;
+	virtual uint8_t Id() const = 0;
 
 	/**
 	 * @brief Get the application owning the working mode
@@ -125,7 +125,7 @@ public:
 	 * @param dest_awm_id Destination working mode ID
 	 * @return A pointer to the TransitionOverheads object
 	 */
-	virtual OverheadPtr_t OverheadInfo(uint16_t dest_awm_id) const = 0;
+	virtual OverheadPtr_t OverheadInfo(uint8_t dest_awm_id) const = 0;
 
 	/**
 	 * @brief Bind resource usages to system resources descriptors
